refactor(vt-orchestrator): split transform and vt stage responsibilities
