--- packages/vt-orchestrator/src/transform/createTransformByBandHandler/handler.ts
@@ -1,7 +1,0 @@
-import type { TransformByBandStageContext } from '~/contexts';
-import type { StageHandler, TransformByBandTaskInput } from '~/types/types';
-import { createTransformByBandHandler as createTransformByBandHandlerImpl } from './execute.js';
-
-export const createTransformByBandHandler = (
-  context: TransformByBandStageContext
-): StageHandler<TransformByBandTaskInput> => createTransformByBandHandlerImpl(context);
