+++ packages/vt-orchestrator/src/transform/createTransformByBandHandler/transformByBandTaskFinalize.ts
@@ -1,7 +1,12 @@
 import type { NodeId } from '@hierarchidb/core-types';
 import type { EphemeralDB } from '@hierarchidb/gis-sdk';
 import { logDebug } from '~/debug/persistentDebugLog';
-import { TASKDEBUG_BUILD_TAG, TRANSFORM_CACHE_WRITE_SLOW_LOG_MS, TRANSFORM_DB_WRITE_TIMEOUT_MS, withTimeout } from './helpers.js';
+import {
+  TASKDEBUG_BUILD_TAG,
+  TRANSFORM_CACHE_WRITE_SLOW_LOG_MS,
+  TRANSFORM_DB_WRITE_TIMEOUT_MS,
+  withTimeout,
+} from './helpers/core.js';
 
 type UpdateTaskStrict = (taskId: string, updates: Record<string, unknown>, operation: string) => Promise<void>;
