+++ packages/vt-orchestrator/src/vt/vtStageHandler.ts
@@ -1,6 +1,6 @@
 import type { VTStageContext } from '~/contexts';
 import type { StageHandler, StageHandlerResult, VtTaskInput } from '~/types/types';
-import { runVtTaskProcessorFlow } from './vtStageTaskProcessorFlow.js';
+import { runVtTaskProcessorFlow } from './vtStageTaskProcessor.js';
 
 export const createVtHandler = (context: VTStageContext): StageHandler<VtTaskInput> => {
   return async (task): Promise<StageHandlerResult> => runVtTaskProcessorFlow({
